Fix: class_exists() error when using stdclass with required interfaces
You could quite happily create a mock for a stdClass with required interfaces
however whenever you tried to enforce and requirements you'd be met with

    class_exists() expects parameter 1 to be string, array given

Example:

    $mock = Mockery::mock('stdClass, ArrayAccess');
    $mock->shouldReceive('foo')->once();
    $mock->foo();

Added new test to check for this exact situation, and fix to `Generator.php`.

In the two places causing error we're either checking for the array and
returning early before the error occurs, or joining the array on a comma to
display the fictional class name in the same format it's written in the
source.

This may or may not be correct as I'm not familiar with Mockery's source, but it doesn't appear to break any other tests.