Default to eval loader
The require loader is great for debugging, but causes a problem in the tests.
When running a test in a separate process, PHPUnit now tries to include all
files that have been included in this process in a new process.

We test MongoCollection with an altered signature and use the @ operator to
suppress the E_STRICT, when loading the mock. We can't control this in the
separate process, so they can always fail, for unrelated code.