Add Hamcrest assertThat() so assertions can be counted
I've integrated Hamcrest with my PHPUnit in RunBare method https://gist.github.com/dharkness/3965902

And when i use Mockery with Hamcrest Matchers as arguments in method f.ex. with() than PHPUnit doesnt count assertion.

Example:
$adapter = Mockery::mock('Logger');
$adapter->shouldReceive('write')->once()->with(hasKey('message')->andReturnNull();

This update pass the matcher to the asserThat method of Hamcrest Framework so the assertions are counted.