Raise a descriptive error when a required class is missing
Currently when Mockery has been configured not to allow mocking
non-existent methods, attempting to mock a non-existent class gives a
confusing error. It complains that it can't mock methods that don't
exist on the specified class, but actually the error is that mockery
can't find the referenced class to mock it.

This commit changes that behaviour so that attempting to mock a class
that doesn't exist throws an exception that explicitly says that the
specified class doesn't exist, however it only does so while mockery is
configured not to allow mocking of non-existent methods.