Restore the PHPUnit listener
The listener has been removed in ede143167ae34e because it is called too
late in the test lifecycle. This marks tests incorrectly as risky, if
there are no other assertions in the test.

Simply removing the listener silently breaks existing testsuites for
PHPUnit <5.5.6, which will not complain about missing listeners.

The listener added in this commit will no longer verify mock assertions.
Instead, the listener will mark tests as failed if they don't verify the
mock assertions themselves.

This will surface the change for existing users of the listener and help them
migrate to the trait or base class. Is also serves as an opt-in safefail for
new Mockery users.

Fixes #636