Add explicit pattern matcher
Stop trying to do implicit regex matching when matching
string arguments, as it tends to not play nicely with
JSON strings as arguments - PHP thinks it's a regex pattern
to match against.

Introduce \Mockery::pattern() and the Pattern matcher to do
explicit regex pattern matching when it's required by the user.

Fixes #338