Adds support for late static binding
Objects that are mocked and then "recreated" with
late static binding would not inherit the expectations set
on the original mock object.

By telling Mockery to consider those mock objects as instance
mocks as well, they will inherit the expectations as expected.

Went with a new prefix of `lsb:` instead of trying to bolt it
on top of existing functionality because I'm too scared of the
things that might break otherwise.