Support PHPUnit 5.7.27
In order to work with older PHPUnit versions (<6), Mockery provides
some cool Legacy support adapters. The TestListenerTrait for example
creates class aliasses in order to autoload the used classes the same
way between PHPUnit versions.

PHPUnit also provides some forwards compatibilty support in the later
5.7 releases. These support features are not fully in sync and will
cause a class not found exception when running tests on older PHP
versions. The least breaking fix is the one I'm suggesting in this
commit. It should work for older PHPUnit versions and also for the
latest 5.7.27.

https://github.com/sebastianbergmann/phpunit/tree/5.7.27/src/ForwardCompatibility