Split the code by the first curly bracket
Given that the str_replace has to happen on substrings that are before
the first curly bracket, split the generated mock code by the first
curly bracket, and then do the str_replace only on that smaller chunk of
code.

Not sure if there's really any benefit to this, as we're replacing one
string operation with another one.