Add PHPUnit 10+ attributes
Fixes #1359

While upgrading to PHPUnit 11 I noticed the following deprecation:

Metadata in doc-comments is deprecated and will no longer be supported in PHPUnit 12.
Update your test code to use attributes instead.

If we add the attribute too, it stops complaining. So this most be a way to support PHPUnit 11
and older versions.

If we ship this, we reduce a few errors. Most of the things just work fine.